@RPGCore Changes 4.0.19 - Dungeons and Bugfixes
* Added instanced dungeons
* Added weapon scaling attributes to apply custom damage multipliers to weapons
* Added "RAGE_CAP_FROM_WEAPON" attribute, to cap weapon rage
* Added "{LEVEL}" placeholder to creature names
* Added skill selector for entity level
* Added a short buffer before players will die
* Added "score" selector and mechanic
* Updated the "/rpg tool" command and the editing process
* Updated minimum movement speed to be 0.01
* Fixed a bug where the generated minimap could overlap
* Fixed a bug where removed mobs would permanently be kept in the entity list
* Fixed a bug where failure to add entities would not give proper blame
* Fixed a bug where sometimes the editor would fail to enforce lower-case text
* Fixed a bug where resourcepack would not re-generate with bad URL
* Fixed a bug where rage reset incorrectly
* Fixed a bug where NPC would not apply items
* Fixed a bug where level based player attributes were incorrectly calculcated
* Fixed a bug where the required/forbidden option broke quests
* Fixed a bug where completed quests could be accepted again
* Fixed a bug where quest delivery/dialogue could fail without traits
* Fixed a bug where non-targetable entities could take damage
* Fixed a bug where player locations did not save/load correctly
* Fixed a bug where the audio effect did not play properly
* Fixed a bug where item rewards from quests could not edit amount
* Fixed a bug where mob leash would not apply at a certain distance
* Fixed a bug where players did not properly respawn
* Fixed a bug where edit highlighting of nodes was shown to others.

+++ plugin/src/main/java/me/blutkrone/rpgcore/hud/editor/bundle/quest/reward/EditorQuestRewardItem.java
@@ -21,7 +21,7 @@ public class EditorQuestRewardItem extends AbstractEditorQuestReward {
     public String item = "NOTHINGNESS";
     @EditorNumber(name = "Amount", minimum = 0, maximum = 64)
     @EditorTooltip(tooltip = "Amount is limited by effective stack size.")
-    public int amount = 1;
+    public double amount = 1;
 
     public EditorQuestRewardItem() {
     }
